core_pb
Module constants

Constants

DEFAULT_NETWORK

The default network the robot tries to connect to

GAME_SERVER_MAGIC_NUMBER

this message lets game server clients know that a game server supports extra messages like pause, reset, custom game state

GAME_SERVER_PORT

The default port where server_pb should expect to find the game server

GUI_LISTENER_PORT

The default port where gui_pb should expect to connect to server_pb

GU_PER_INCH

Grid units per inch

GU_PER_M

Grid units per meter

INCHES_PER_GU

Inches per grid unit

INCHES_PER_M

Inches per meter

INITIAL_DISPLAY_LOOP_INTERVAL

INITIAL_ENABLE_BATTERY_MONITOR

INITIAL_ENABLE_DISPLAY

INITIAL_ENABLE_DISTS

INITIAL_ENABLE_EXTRA_IMU_DATA

INITIAL_ENABLE_GAMEPAD

INITIAL_ENABLE_IMU

MAX_ROBOT_PATH_LENGTH

The maximum number of nodes in the target path sent from the server to the robot

MM_PER_GU

Millimeters per grid unit

MM_PER_INCH

Millimeters per inch

PWM_SOFT_CAP

Hardcoded maximum PWM signal that the pico will send to motors, as a safety

ROBOT_DISPLAY_HEIGHT

The size of the OLED display on the robot

ROBOT_DISPLAY_WIDTH

The size of the OLED display on the robot

ROBOT_LOGS_BUFFER

SIMULATION_LISTENER_PORT

The default port where server_pb should expect to find the simulation controls

SOCKET_TIMEOUT

The default timeout period in seconds between socket messages after which a socket attempts to reconnect