core_pb
Module messages

Modules

common

ota

robot_tcp

server_status

settings

Structs

ExtraImuData

ExtraOptsTypes

FrequentServerToRobot

This is sent regularly and frequently to robots via ServerToRobotMessage::FrequentRobotItems

MotorControlStatus

RobotStatus

SensorData

Sent from the robot peripherals task to the wifi task and back to the server

Enums

GameServerCommand

GuiToServerMessage

Messages sent from gui_pb to server_pb

NetworkStatus

RobotButton

RobotToServerMessage

Firmware related items MUST remain first, or OTA programming will break

ServerToGuiMessage

Messages sent from server_pb to gui_pb

ServerToRobotMessage

Firmware related items MUST remain first, or OTA programming will break

ServerToSimulationMessage

SimulationToServerMessage

Messages sent from sim_pb to server_pb

Task

The different async tasks that run on the robot

VelocityControl

Constants

MAX_SENSOR_ERR_LEN

Type Aliases

ExtraOptsAtomicTypes