core_pb
Module region_localization

Structs

Region

A Region of a Grid is an area where moving around the region yields continuous theoretical distance sensor readings

Constants

VECTORS 🔒

Functions

estimate_location

estimate_location_2

get_at 🔒

get_possible_regions

get_region_score 🔒