Module core_pb::threaded_websocket

See ThreadedSocket, a simple poll-based wrapper around a socket (websocket or TCP) connection that runs in a separate thread

Structs

TcpStreamThreadableSocket

A TCP socket compatible with ThreadedSocket

ThreadedSocket

Simple poll-based wrapper around a socket (websocket or TCP) connection that runs in a separate thread

Enums

TextOrT

Represents data that is either the given type, or text

Traits

ThreadableSocket

Represents a type that is compatible with ThreadedSocket

Functions

run_socket_forever 🔒

Runs on a separate thread to babysit the socket

socket_read_fut 🔒

A future that yields the next message from the socket, or never if the socket is None

Type Aliases

Address

ipv4 address with port number